+++ pkg/yamlParser/yamlparser.go
@@ -5,6 +5,7 @@ import (
 	"os"
 
 	yaml "github.com/goccy/go-yaml"
+	"github.com/xaaha/hulak/pkg/actions"
 	"github.com/xaaha/hulak/pkg/envparser"
 	"github.com/xaaha/hulak/pkg/utils"
 )
@@ -86,7 +87,7 @@ func checkYamlFile(filepath string, secretsMap map[string]any) (*bytes.Buffer, e
 	parsedMap := replaceVarsWithValues(data, secretsMap)
 
 	// translate the types, if acceptable
-	parsedMap, err = translateType(data, parsedMap, secretsMap, envparser.GetValueOf)
+	parsedMap, err = translateType(data, parsedMap, secretsMap, actions.GetValueOf)
 	if err != nil {
 		return nil, utils.ColorError("#reader", err)
 	}
@@ -101,8 +102,9 @@ func checkYamlFile(filepath string, secretsMap map[string]any) (*bytes.Buffer, e
 	return &buf, nil
 }
 
-// checks the validity of all the fields in the yaml file meant for regular api call
-func FinalStructForApi(filePath string, secretsMap map[string]any) (ApiCallFile, error) {
+// FinalStructForAPI builds a final struct for the api call.
+// It  checks the validity of all the fields in the yaml file meant for regular api call
+func FinalStructForAPI(filePath string, secretsMap map[string]any) (ApiCallFile, error) {
 	buf, err := checkYamlFile(filePath, secretsMap)
 	if err != nil {
 		return ApiCallFile{}, utils.ColorError("Error occured after reading yaml file", err)
